requirements.txt
ajsonrpc==1.2.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.69
botocore==1.40.69
bottle==0.13.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
dnspython==2.8.0
fastapi==0.121.1
h11==0.16.0
idna==3.10
jmespath==1.0.1
marshmallow==3.26.1
packaging==25.0
platformio==6.1.18
pydantic==2.12.4
pydantic_core==2.41.5
pyelftools==0.32
pymongo==4.15.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
s3transfer==0.14.0
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.34.3
wsproto==1.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
authlib==1.3.0
pydantic[email]>=2.0.0
requests==2.32.5
httpx
fastapi-mail
itsdangerous
